refactor emp migrated and welcome bare bones
